+++ src/components/ui/VideoTile/index.tsx
@@ -1,0 +1,40 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import React, { forwardRef, HTMLAttributes } from 'react';
+
+import { StyledVideoTile } from './Styled';
+import { BaseProps } from '../Base';
+
+type ObjectFit = 'contain' | 'cover' | 'fill' | 'none' | 'scale-down';
+
+export interface VideoTileProps
+  extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>,
+    BaseProps {
+  nameplate?: string;
+  objectFit?: ObjectFit;
+}
+
+export const VideoTile = forwardRef(
+  (props: VideoTileProps, ref: React.Ref<HTMLVideoElement>) => {
+    const { tag, className, nameplate, ...rest } = props;
+
+    return (
+      <StyledVideoTile
+        as={tag}
+        className={className || ''}
+        data-testid="video-tile"
+        {...rest}
+      >
+        <video ref={ref} className="video" />
+        {nameplate && (
+          <header className="nameplate">
+            <p className="text">{nameplate}</p>
+          </header>
+        )}
+      </StyledVideoTile>
+    );
+  }
+);
+
+export default VideoTile;