+++ tst/components/ui/PopOver/PopOverSubMenu.test.tsx
@@ -1,0 +1,52 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import '@testing-library/jest-dom';
+import React from 'react';
+
+import PopOverSubMenu from '../../../../src/components/ui/PopOver/PopOverSubMenu';
+import lightTheme from '../../../../src/theme/light';
+import { renderWithTheme } from '../../../test-helpers';
+import PopOverItem from '../../../../src/components/ui/PopOver/PopOverItem';
+import { fireEvent } from '@testing-library/dom';
+
+describe('PopOver submenu', () => {
+
+  it('should render a popover submenu component', () => {
+    const component = <PopOverSubMenu text='submenu'/>;
+    const { getByText } = renderWithTheme(lightTheme, component);
+    const element = getByText('submenu');
+
+    expect(element).toBeInTheDocument();
+  });
+
+  it('should open a submenu of options when clicked', () => {
+    const component = <PopOverSubMenu text='submenu' children={<PopOverItem children={<span>Option 1</span>}/>}/>;
+    const { getByText } = renderWithTheme(lightTheme, component);
+    const submenu = getByText('submenu');
+    fireEvent.click(submenu);
+    const option1 = getByText('Option 1')
+
+    expect(option1).toBeInTheDocument();
+    fireEvent.click(submenu);
+  });
+
+  it('should close the new popover when clicked twice', () => {
+    const component = <PopOverSubMenu text='submenu' children={<PopOverItem children={<span>Option 1</span>}/>}/>;
+    const { getByText } = renderWithTheme(lightTheme, component);
+    const submenu = getByText('submenu');
+    fireEvent.click(submenu);
+    const option1 = getByText('Option 1');
+    fireEvent.click(submenu);
+
+    expect(option1).not.toBeInTheDocument();
+  });
+
+  it('should render a caret', () => {
+    const component = <PopOverSubMenu text='submenu'/>;
+    const { getByTestId } = renderWithTheme(lightTheme, component);
+    const caret = getByTestId('submenu-caret');
+
+    expect(caret).toBeInTheDocument();
+  });
+});